interface is a point-to-point link
Interfaces in HALO container represent point-to-point links
between neighbors. For example, in a setup with 3 routers
each connected to the network via one physical link, HALO
containers have two interfaces each:

   +--------+   +--------+   +--------+
   + HALO-A |   + HALO-B |   + HALO-C |
   +-+----+-+   +-+----+-+   +-+----+-+
     |AB  |AC     |BA  |BC     |CA  |CB
     |    |       |    |       |    |
   +-+----+-+   +-+----+-+   +-+----+-+
   + DNOS-A |   + DNOS-B |   + DNOS-C |
   +---+----+   +----+---+   +----+---+
       |             |            |
     ==+=============+============+==

Delay and jitter in this case are attributes of the interface
and they represent values measured for the corresponding link
(or tunnel) between current HALO container and its peer.

Configure flow steering by setting egress interface instead
of next-hop IP. This means switching to the corresponding
point-to-point link (or tunnel) when sending out packets.